refactor(ios): split settings dashboard helpers
